Increase default password size to 10
Some people recommend longer passwords, but I feel if you really need a
longer password than 10, you should switch to using several words that
are easier to remember. At any rate, this is the kind of setting that
may need to go up as time passes, even if it breaks backward
compatibility.